pyproject.toml
[project]
name = "fast-api-test"
version = "0.1.0"
description = ""
authors = [
    {name = "Dimas Naufal",email = "dimasnaufalxyz@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types (==0.7.0)",
    "anyio (==4.9.0)",
    "cffi (==1.17.1)",
    "click (==8.1.8)",
    "colorama (==0.4.6)",
    "cryptography (==44.0.2)",
    "fastapi (==0.115.12)",
    "greenlet (==3.1.1)",
    "h11 (==0.14.0)",
    "idna (==3.10)",
    "passlib (==1.7.4)",
    "pycparser (==2.22)",
    "pydantic (==2.10.6)",
    "pydantic-core (==2.27.2)",
    "pyjwt (==2.10.1)",
    "pymysql (==1.1.1)",
    "python-multipart (==0.0.20)",
    "redis (==5.2.1)",
    "setuptools (==78.1.0)",
    "sniffio (==1.3.1)",
    "sqlalchemy (==2.0.40)",
    "starlette (==0.46.1)",
    "types-cffi (==1.17.0.20250326)",
    "types-pyopenssl (==24.1.0.20240722)",
    "types-redis (==4.6.0.20241004)",
    "types-setuptools (==76.0.0.20250313)",
    "typing-extensions (==4.13.0)",
    "uvicorn (==0.34.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
